make get_stats return float

+++ optd-persistent/src/orm_manager.rs
@@ -45,7 +45,7 @@ impl StorageLayer for ORMManager {
         table_id: i32,
         stat_type: i32,
         epoch_id: Option<EpochId>,
-    ) -> StorageResult<Option<i32>> {
+    ) -> StorageResult<Option<f32>> {
         todo!()
     }
 
@@ -54,7 +54,7 @@ impl StorageLayer for ORMManager {
         attr_id: i32,
         stat_type: i32,
         epoch_id: Option<EpochId>,
-    ) -> StorageResult<Option<i32>> {
+    ) -> StorageResult<Option<f32>> {
         todo!()
     }
 
@@ -63,7 +63,7 @@ impl StorageLayer for ORMManager {
         attr_ids: Vec<i32>,
         stat_type: i32,
         epoch_id: Option<EpochId>,
-    ) -> StorageResult<Option<i32>> {
+    ) -> StorageResult<Option<f32>> {
         todo!()
     }
